api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==0.28.1
python-dotenv==1.0.0
PyPDF2==3.0.1
sentence-transformers
faiss-cpu
numpy